+++ packages/flutter_adaptive_cards_fs/lib/flutter_adaptive_cards_fs.dart
@@ -13,6 +13,7 @@ export 'package:flutter_adaptive_cards_fs/src/flutter_raw_adaptive_card.dart'
 export 'package:flutter_adaptive_cards_fs/src/hostconfig/host_config.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/action_invoke.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/adaptive_card_update.dart';
+export 'package:flutter_adaptive_cards_fs/src/models/authentication_config.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/choice.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/data_query.dart';
 export 'package:flutter_adaptive_cards_fs/src/models/fact.dart';