+++ packages/flutter_adaptive_cards_host_fs/lib/src/adapters/teams_invoke_adapter.dart
@@ -64,6 +64,14 @@ class TeamsInvokeAdapter {
             },
           },
         };
+      case AdaptiveCardInvokeKind.signin:
+        return {
+          'type': 'invoke',
+          'name': 'signin/verifyState',
+          'value': {
+            'state': request.value?.toString(),
+          },
+        };
     }
   }